asignaTareaUsuario -- con nullpointerexception revisar

+++ SGTI_EJB/ejbModule/stateless/ManagerU.java
@@ -54,7 +54,7 @@ public class ManagerU implements UsuarioRemote{
                 return todos;
         }
         
-        public Usuario encontrarUsuario(int ced){
+        public Usuario encontrarUsuario(long ced){
                 Usuario u = em.find(Usuario.class, ced);
                 return u;
         }
